Pty refactor
Code will be easier to maintain down the road.

Using `net.Socket` to wrap fds by overriding some internal NodeJS
function was hacky too, and also did not work correctly on Mac.

Hopefully the `FileDuplex` wrapper will not suffer the same issues, it
already does not rely on a hack, which is better.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>